docker/collector/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pydantic==2.9.2
prometheus-client==0.21.0
httpx==0.27.2
